Add menu items in admin board

+++ app/models/MenuItem.py
@@ -5,7 +5,7 @@ class MenuItem(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     name = db.Column(db.String(80), unique=True, nullable=False)
     image = db.Column(db.String(255), nullable=False)
-    price = db.Column(db.Integer, nullable=False)
+    price = db.Column(db.Float, nullable=False)
     description = db.Column(db.Text, nullable=False)
     menu_id = db.Column(db.Integer, db.ForeignKey("menu.id"), nullable=False)
